* Remove documentation in the uninstall target * Add 'autogen.sh' to ease the use of 'acr'   - Added USERCC and USEROSTYPE configuration flags     --with-compiler : select a compiler from mk/     --with-ostype   : select a target OS (not required for ming32 f.ex) * Add language bindings APIs from radare1   - Fix some issues from the ruby language binding   - Checks if file exists before slurping and causing a ruby segfault   - Added language destructor (.fini pointer) * Rename str_clean into r_str_chop * Initial work on the port for ming-w32 * Show flags instead of bytes in disassembly (smart disasm) * New r_sys_{get|set}env (for portability issues * Added flags -i and -l to radare2 * Fix warnings and random code cleanup * Added r_flag_get_i

+++ libr/diff/t/test.c
@@ -62,7 +62,7 @@ int test_distance()
 	char *bufa = "hello";
 	char *bufb = "heprpworld";
 	u32 distance = 0;
-	float similarity = 0;
+	double similarity = 0;
 
 	printf("Similarity: '%s' vs '%s'\n", bufa, bufb);
 	r_diff_buffers_distance(NULL, bufa, strlen(bufa), bufb, strlen(bufb),
